Refactor NPIE codebase for improved functionality and maintainability
- Updated logging function to correctly format microseconds in timestamps.
- Removed unused source files from CMakeLists.txt to streamline build process.
- Enhanced model metadata structure to include model size and quantization status.
- Modified emlearn backend load function to accept model buffer and size parameters.
- Cleaned up npie_core.c by removing commented-out code and unnecessary whitespace.
- Implemented new structures for scheduler and memory manager statistics.
- Updated accelerator initialization and cleanup functions to use new accelerator descriptor structure.
- Improved error handling and logging for accelerator operations.
- Refactored model loading functions to better manage memory and model data.
- Added support for detecting and managing hardware accelerators in the HAL.

+++ src/libs/libneura_common.c
@@ -73,8 +73,8 @@ void neura_log(neura_log_level_t level, const char* format, ...) {
         struct timeval tv;
         gettimeofday(&tv, NULL);
         
-        fprintf(stderr, "[%ld.%06d] [%s] %s\n",
-                (long)tv.tv_sec, tv.tv_usec, level_str[level], buffer);
+        fprintf(stderr, "[%ld.%06ld] [%s] %s\n",
+                (long)tv.tv_sec, (long)tv.tv_usec, level_str[level], buffer);
     }
 }
